split into cards
