+++ resources/js/Pages/Shipments/Partials/ShipmentFinancialDetails.tsx
@@ -1,6 +1,4 @@
 import AccessorialsTable from '@/Components/Shipments/AccessorialsTable';
-import CarrierRatesTable from '@/Components/Shipments/CarrierRatesTable';
-import CustomerRatesTable from '@/Components/Shipments/CustomerRatesTable';
 import { Card, CardContent, CardHeader, CardTitle } from '@/Components/ui/card';
 import { Loading } from '@/Components/ui/loading';
 import {
@@ -13,6 +11,8 @@ import {
 import { BadgeDollarSign, Users } from 'lucide-react';
 import { useEffect, useState } from 'react';
 import CustomerRates from './CustomerRates';
+import CarrierRates from '@/Pages/Shipments/Partials/CarrierRates';
+import Accessorials from '@/Pages/Shipments/Partials/Accessorials';
 
 export default function ShipmentFinancialDetails({
     shipment,
@@ -96,51 +96,9 @@ export default function ShipmentFinancialDetails({
         <>
             <CustomerRates shipment={shipment} shipmentFinancials={shipmentFinancials}/>
             
-            <Card>
-                <CardHeader>
-                    <CardTitle className="flex items-center gap-2">
-                        <BadgeDollarSign className="h-5 w-5" />
-                        Carrier Rates
-                    </CardTitle>
-                </CardHeader>
-                <CardContent>
-                    {isLoading ? (
-                        <Loading
-                            className="mx-auto h-[200px] w-full"
-                            text="Loading..."
-                        />
-                    ) : (
-                        <CarrierRatesTable
-                            rate_types={carrierRateTypes}
-                            rates={shipmentFinancials?.shipment_carrier_rates ?? []}
-                            shipment={shipment}
-                        />
-                    )}
-                </CardContent>
-            </Card>
+            <CarrierRates shipment={shipment} shipmentFinancials={shipmentFinancials}/>
             
-            <Card>
-                <CardHeader>
-                    <CardTitle className="flex items-center gap-2">
-                        <BadgeDollarSign className="h-5 w-5" />
-                        Accessorials
-                    </CardTitle>
-                </CardHeader>
-                <CardContent>
-                    {isLoading ? (
-                        <Loading
-                            className="mx-auto h-[200px] w-full"
-                            text="Loading..."
-                        />
-                    ) : (
-                        <AccessorialsTable
-                            accessorial_types={accessorialTypes}
-                            accessorials={shipmentFinancials?.accessorials ?? []}
-                            shipment={shipment}
-                        />
-                    )}
-                </CardContent>
-            </Card>
+            <Accessorials shipment={shipment} shipmentFinancials={shipmentFinancials}/>
         </>
     );
 }
